Move fixHref and fixAs to first level
